requirements.txt
aws-cdk.core==1.125.0
aws-cdk.aws-s3==1.125.0
aws-cdk.aws-lambda==1.125.0
aws-cdk.aws-sqs==1.125.0
aws-cdk.aws-dynamodb==1.125.0
aws-cdk.aws-apigateway==1.125.0
aws-cdk.aws-lambda-event-sources==1.125.0
